+++ src/main/resources/db/changelog/changes/022-create-product-bundle-offer.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS product_bundle_special_offer_product (
-     product_bundle_id INT,
-     special_offer_product_id INT,
+     product_bundle_id BIGINT,
+     special_offer_product_id BIGINT,
      PRIMARY KEY (product_bundle_id, special_offer_product_id),
      FOREIGN KEY (product_bundle_id) REFERENCES products_bundle(id),
      FOREIGN KEY (special_offer_product_id) REFERENCES special_offer_product(id)
