+++ src/main/resources/db/changelog/changes/033-create-product-bundle-options.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS product_bundle_options (
-      id INT AUTO_INCREMENT PRIMARY KEY,
-      product_bundle_id INT NOT NULL,
+      id BIGINT AUTO_INCREMENT PRIMARY KEY,
+      product_bundle_id BIGINT NOT NULL,
       quantity INT NOT NULL,
       price DECIMAL(7,2) NOT NULL,
       FOREIGN KEY (product_bundle_id) REFERENCES products_bundle(id) ON DELETE CASCADE
